chore: revert changes, make rockcraft-specific
