+++ craft_platforms/_build.py
@@ -17,7 +17,7 @@
 
 from typing import Any, Callable, Dict, Iterable
 
-from craft_platforms import _errors, charm, rock, snap
+from craft_platforms import charm, rock, snap
 from craft_platforms._buildinfo import BuildInfo
 from craft_platforms._platforms import get_platforms_build_plan
 
@@ -54,10 +54,6 @@ def get_build_plan(
         "platforms": project_data.get("platforms"),
         "build_base": project_data.get("build-base"),
     }
-
-    # Bare bases require a build_base
-    if args["base"] == "bare" and args["build_base"] is None:
-        raise _errors.NeedBuildBaseError(base=args["base"])
 
     if app == "snapcraft":
         args["snap_type"] = project_data.get("type")
