+++ craft_platforms/rock/_build.py
@@ -33,7 +33,12 @@ def get_rock_build_plan(
     :param base: the rock base (e.g. ``'ubuntu@24.04'``)
     :param platforms: the platforms structure in ``rockcraft.yaml``
     :param build_base: the build base, if provided in ``rockcraft.yaml``.
+    :raises NeedsBuildBaseError: If base is bare and build_base is empty
     """
+    # Bare bases require a build_base
+    if base == "bare" and build_base is None:
+        raise _errors.NeedBuildBaseError(base=base)
+
     for name, platform in platforms.items():
         if platform and "all" in platform.get("build-for", []):
             raise _errors.InvalidPlatformError(
