+++ docs/reference/changelog.rst
@@ -7,8 +7,8 @@ Changelog
 Bug Fixes
 =========
 
-- The ``bare`` base for build plans was not properly handled. An error will now be
-  raised if ``bare`` is specified with no ``build-base``.
+- The ``bare`` base for Rockcraft was not properly handled. An error will now be raised
+  if ``bare`` is specified with no ``build-base``.
 
 0.7.0 (2025-04-02)
 ------------------
